Change tank to a dual robot game.
6/6/2016 9:45AM release
Change the setup to put the robot tanks in a line in front of the battle
tanks and make the red a robot. Now we have a fair fight. Done.
Separate the red robot and yellow robot tank logic for future
competition. Done
Add an attribute to the robot tanks call tank_type. Either 'Scout' or
'Master'. Done